requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pydantic==2.5.1
jinja2==3.1.2
python-multipart==0.0.6
python-Levenshtein==0.23.0
